# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentForge API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/finetune": {
      "get": {
        "tags": [
          "finetune"
        ],
        "summary": "List Finetune Jobs",
        "operationId": "list_finetune_jobs_api_v1_finetune_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FinetuneJobResponse"
                  },
                  "type": "array",
                  "title": "Response List Finetune Jobs Api V1 Finetune Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Create Finetune Job",
        "operationId": "create_finetune_job_api_v1_finetune_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinetuneCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinetuneJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/finetune/trigger": {
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Trigger Auto Finetune",
        "operationId": "trigger_auto_finetune_api_v1_finetune_trigger_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinetuneTriggerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinetuneJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/finetune/deployed": {
      "get": {
        "tags": [
          "finetune"
        ],
        "summary": "List Deployed Models",
        "description": "List completed fine-tune jobs that have inference endpoints (usable as agent providers).",
        "operationId": "list_deployed_models_api_v1_finetune_deployed_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FinetuneJobResponse"
                  },
                  "type": "array",
                  "title": "Response List Deployed Models Api V1 Finetune Deployed Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/finetune/{job_id}": {
      "get": {
        "tags": [
          "finetune"
        ],
        "summary": "Get Finetune Job",
        "operationId": "get_finetune_job_api_v1_finetune__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinetuneJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "finetune"
        ],
        "summary": "Delete Finetune Job",
        "operationId": "delete_finetune_job_api_v1_finetune__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/{job_id}/deploy": {
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Deploy Finetune",
        "operationId": "deploy_finetune_api_v1_finetune__job_id__deploy_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinetuneJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/{job_id}/cancel": {
      "delete": {
        "tags": [
          "finetune"
        ],
        "summary": "Cancel Finetune Job",
        "operationId": "cancel_finetune_job_api_v1_finetune__job_id__cancel_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/{job_id}/evaluate": {
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Evaluate Finetune",
        "description": "Run evaluation prompts against a deployed fine-tuned model.",
        "operationId": "evaluate_finetune_api_v1_finetune__job_id__evaluate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evaluate Finetune Api V1 Finetune  Job Id  Evaluate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/{job_id}/stream": {
      "get": {
        "tags": [
          "finetune"
        ],
        "summary": "Stream Finetune Job",
        "description": "SSE stream for a fine-tuning job. Subscribes to Redis Pub/Sub channel finetune:{job_id}.",
        "operationId": "stream_finetune_job_api_v1_finetune__job_id__stream_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/{job_id}/inference-stream": {
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Inference Stream",
        "description": "Proxy streaming tokens from Modal inference endpoint.",
        "operationId": "inference_stream_api_v1_finetune__job_id__inference_stream_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceStreamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/finetune/datasets/upload": {
      "post": {
        "tags": [
          "finetune"
        ],
        "summary": "Upload Dataset",
        "description": "Upload a dataset file (JSON, JSONL, CSV, Parquet) for fine-tuning.",
        "operationId": "upload_dataset_api_v1_finetune_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_finetune_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Upload Dataset Api V1 Finetune Datasets Upload Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/finetune/datasets": {
      "get": {
        "tags": [
          "finetune"
        ],
        "summary": "List Datasets",
        "description": "List uploaded datasets for the current user.",
        "operationId": "list_datasets_api_v1_finetune_datasets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Datasets Api V1 Finetune Datasets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/finetune/datasets/{filename}": {
      "delete": {
        "tags": [
          "finetune"
        ],
        "summary": "Delete Dataset",
        "description": "Delete an uploaded dataset.",
        "operationId": "delete_dataset_api_v1_finetune_datasets__filename__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_finetune_datasets_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_finetune_datasets_upload_post"
      },
...
      "FinetuneCreateRequest": {
        "properties": {
          "base_model": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base Model"
          },
          "modality": {
            "type": "string",
            "maxLength": 32,
            "title": "Modality",
            "description": "Training recipe. Only text_sft (Unsloth causal LM) is implemented.",
            "default": "text_sft"
          },
          "dataset_path": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Dataset Path",
            "description": "Hub: hf://org/dataset or hf://org/dataset/config (e.g. hf://openai/gsm8k/main). Multi-config datasets auto-pick 'main' when omitted."
          },
          "hyperparams": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hyperparams"
          }
        },
        "type": "object",
        "required": [
          "base_model",
          "dataset_path"
        ],
        "title": "FinetuneCreateRequest"
      },
      "FinetuneJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model"
          },
          "modality": {
            "type": "string",
            "title": "Modality"
          },
          "dataset_path": {
            "type": "string",
            "title": "Dataset Path"
          },
          "hyperparams": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hyperparams"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "modal_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modal Job Id"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "model_output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Output Path"
          },
          "inference_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inference Endpoint"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "base_model",
          "modality",
          "dataset_path",
          "hyperparams",
          "status",
          "modal_job_id",
          "metrics",
          "model_output_path",
          "inference_endpoint",
          "started_at",
          "completed_at",
          "created_at"
        ],
        "title": "FinetuneJobResponse"
      },
      "FinetuneTriggerRequest": {
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "title": "Agent Id"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "default": "unsloth/llama-3-8b-Instruct"
          },
          "min_score": {
            "type": "number",
            "title": "Min Score",
            "default": 0.8
          }
        },
        "type": "object",
        "required": [
          "agent_id"
        ],
        "title": "FinetuneTriggerRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InferenceStreamRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "max_new_tokens": {
            "type": "integer",
            "title": "Max New Tokens",
            "default": 128
          },
          "temperature": {
            "type": "number",
            "title": "Temperature",
            "default": 0.7
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "InferenceStreamRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}